add directory models, update tests
