+++ .coverage
@@ -1,0 +1,1 @@
+!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\USA\\Desktop\\andela35\\iReporterApi\\tests\\__init__.py":[1],"C:\\Users\\USA\\Desktop\\andela35\\iReporterApi\\tests\\test_app.py":[1,2,3,6,7,38,41,50,61,67,72,76,84,99,107,115,121,11,13,14,15,16,17,20,21,22,23,24,27,28,29,30,31,32,33,35,36,51,52,53,54,56,59,39,85,86,87,88,89,90,91,93,94,97,100,101,102,103,104,105,116,117,118,119,122,123,124,125,126,127,62,63,64,65,68,69,70,73,74,77,78,79,80,81,82,45,46,47,48,108,109,110,111,112],"C:\\Users\\USA\\Desktop\\andela35\\iReporterApi\\app\\__init__.py":[1],"C:\\Users\\USA\\Desktop\\andela35\\iReporterApi\\app\\views.py":[1,2,3,4,7,9,12,21,46,59,83,94,106,117,132,64,65,66,67,68,69,70,71,72,74,75,77,78,79,111,112,113,114,122,123,124,125,126,128,129,88,89,90,100,101,103,102,17,18,73,26,27,28,29,30,31,32,33,34,35,36,38,39,40,41,42,43,51,52,54],"C:\\Users\\USA\\Desktop\\andela35\\iReporterApi\\app\\models\\__init__.py":[1],"C:\\Users\\USA\\Desktop\\andela35\\iReporterApi\\app\\models\\incident.py":[1,3,6,7,9,19,34,43,51,61,71,10,11,12,13,14,15,16,17,39,24,25,26,27,28,29,30,31,40,41,66,67,68,69,77,78,80,79,48,49,57,58,59],"C:\\Users\\USA\\Desktop\\andela35\\iReporterApi\\app\\models\\user.py":[1,3,6,9,11,23,41,47,51,12,13,14,15,16,17,18,19,20,21,43,28,29,30,31,32,33,34,35,36,37,44,45,49],"C:\\Users\\USA\\Desktop\\andela35\\iReporterApi\\app\\validator.py":[1,4,7,9,22,26,11,13,15,17,19,16,27,29,31,33,35],"C:\\Users\\USA\\Desktop\\andela35\\iReporterApi\\tests\\test_users.py":[1,2,3,6,7,22,27,35,11,13,14,15,16,17,18,19,36,37,38,39,40,41,42,23,24,25,28,29,30,31,32,33]}}
