+++ .travis.yml
@@ -1,7 +1,7 @@
 language : python
 
 version :
-  - 3.6.5
+  - 3.6
 
 install:
   - pip install -r requirements.txt
@@ -11,3 +11,4 @@ script:
 
 after_script:
   - coveralls
+
