+++ requirements.txt
@@ -10,6 +10,7 @@ coverage==4.5.2
 coveralls==1.5.1
 docopt==0.6.2
 Flask==1.0.2
+Flask-JWT-Extended==3.15.0
 gunicorn==19.9.0
 idna==2.8
 isort==4.3.4
@@ -22,6 +23,7 @@ more-itertools==4.3.0
 pluggy==0.8.0
 py==1.7.0
 pycodestyle==2.4.0
+PyJWT==1.7.1
 pylint==1.9.3
 pytest==4.0.2
 pytest-cov==2.6.0
